backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
requests
httpx
openai
passlib[bcrypt]
python-jose
pymysql
cryptography
bcrypt==4.0.1
authlib
pip freeze 